スキーマ情報更新

+++ sql/schema/tbl_ptnum.sql
@@ -1,17 +1,17 @@
 CREATE TABLE tbl_ptnum (
-    ptid numeric(10,0) NOT NULL,
+    ptid bigint NOT NULL,
     ptnum character(20) NOT NULL,
-    hknid numeric(10,0) DEFAULT 0,
-    kohid numeric(10,0) DEFAULT 0,
-    autocombinum numeric(4,0) DEFAULT 0,
-    manucombinum numeric(4,0) DEFAULT 0,
+    hknid bigint DEFAULT 0,
+    kohid bigint DEFAULT 0,
+    autocombinum smallint DEFAULT 0,
+    manucombinum smallint DEFAULT 0,
     tstptnumkbn character(1),
     termid character varying(16),
     opid character varying(16),
     creymd character(8),
     upymd character(8),
     uphms character(6),
-    hospnum numeric(2,0) NOT NULL
+    hospnum smallint NOT NULL
 );
 
 COMMENT ON TABLE tbl_ptnum IS '患者番号変換';
