+++ sql/schema/view_bd001.sql
@@ -12,8 +12,8 @@ CREATE VIEW view_bd001 AS
       FROM tbl_syunou_main a,
            tbl_ptnum b,
            tbl_ptinf c 
-     WHERE (((((((((a.hospnum)::numeric = b.hospnum) 
-       AND ((a.ptid)::numeric = b.ptid)) 
+     WHERE ((((((((a.hospnum = b.hospnum) 
+       AND (a.ptid = b.ptid)) 
        AND ((a.hospnum)::numeric = c.hospnum)) 
        AND ((a.ptid)::numeric = c.ptid)) 
        AND (a.denpjtikbn <> '3'::bpchar)) 
@@ -21,6 +21,6 @@ CREATE VIEW view_bd001 AS
        AND (a.createkbn <> '3'::bpchar)) 
        AND (c.tstptnumkbn <> '1'::bpchar));
 
-COMMENT 
-        ON VIEW view_bd001 IS '¼ýÇ¼¥Ó¥å¡¼';
+
+COMMENT ON VIEW view_bd001 IS '¼ýÇ¼¥Ó¥å¡¼';
